ai_chat_portal/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# AI Integration
google-genai==1.48.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
setuptools>=70.0.0

# For embeddings and semantic search
numpy==1.26.2
scikit-learn==1.3.2

# API Documentation
drf-yasg==1.21.7

# Development
black==23.12.1
flake8==6.1.0